requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attr==0.3.1
attrs==23.2.0
autopage==0.5.2
azure-core==1.32.0
azure-storage-blob==12.25.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.12.3
black==24.8.0
bleach==5.0.1
blinker==1.4
block-tracing==1.0.1
boto==2.49.0
boto3==1.37.15
botocore==1.37.15
bottle==0.13.4
boxing==0.1.4
Brlapi==0.7.0
cachetools==5.5.0
certifi==2019.11.28
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cliff==4.7.0
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.1.1
cma==4.0.0
cmaes==0.12.0
cmd2==2.5.11
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
command-not-found==0.3
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
Cython==3.0.11
dask==2023.5.0
dataclasses==0.6
dbus-python==1.2.16
dbxfs==1.0.66
debugpy==1.8.7
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
distro-info===0.23ubuntu1
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
dropbox==11.36.2
duplicity==0.8.12.0
easydict==1.13
efficientnet-pytorch==0.7.1
einops==0.8.0
entrypoints==0.3
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
expiringdict==1.2.2
fasteners==0.14.1
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.3
future==0.18.2
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.2
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
hachoir==3.3.0
hf-xet==1.1.8
htmlmin==0.1.12
httpcore==1.0.7
httplib2==0.14.0
httpx==0.28.1
huggingface-hub==0.34.4
humansignal-drf-yasg==1.21.10.post1
Hydra==2.5
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
idna==2.8
ijson==3.3.0
ImageHash==4.3.2
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
ipykernel==6.29.5
ipympl==0.9.3
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.7
isodate==0.7.2
isort==5.13.2
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json-tricks==3.17.3
jsonschema==3.2.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
keyring==18.0.1
keyrings.alt==4.2.0
kiwisolver==1.4.5
label-studio==1.13.1
label-studio-sdk==1.0.5
language-selector==0.1
launchdarkly-server-sdk==8.2.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.4
lightning-utilities==0.11.2
llvmlite==0.41.1
locket==1.0.0
lockfile==0.12.2
loguru==0.7.2
louis==3.12.0
lxml==5.3.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmcv==1.4.0
mmengine==0.10.5
monotonic==1.5
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
munkres==1.1.4
mypy-extensions==1.0.0
narwhals==1.12.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
ninja==1.11.1.1
nltk==3.9.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
nystrom-attention==0.0.12
oauthlib==3.1.0
olefile==0.46
omegaconf==2.3.0
onnx==1.17.0
onnx-simplifier==0.4.10
openai==1.66.5
opencv-contrib-python==4.12.0.88
opentelemetry-api==1.31.0
optuna==4.5.0
optuna-dashboard==0.19.0
ordered-set==4.0.2
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
partd==1.4.1
pathspec==0.12.1
pbr==7.0.1
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
ply==3.11
prettytable==3.11.0
privy==6.0.0
proglog==0.1.10
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.28.3
psutil==6.1.0
psycopg2-binary==2.9.9
PuLP==3.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.16.2
pycocotools==2.0.7
pycparser==2.22
pycups==1.9.73
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynndescent==0.5.12
pyparsing==3.1.2
pyperclip==1.9.0
pyRFC3339==1.1
pyrsistent==0.20.0
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python-json-logger==2.0.4
pytorch-lightning==2.2.4
pytz==2022.7.1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==6.0.1
pyzmq==26.2.0
redis==3.5.3
referencing==0.35.1
regex==2024.11.6
reportlab==3.5.34
requests==2.32.3
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
rich==13.9.3
rpds-py==0.20.0
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
semver==3.0.4
sentry-sdk==2.33.2
simplejson==3.16.0
six==1.14.0
slack==0.0.2
slack-client==0.3.0
slackclient==2.9.4
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.43
sqlparse==0.5.3
ssh-import-id==5.10
stack-data==0.6.3
stevedore==5.3.0
stone==3.3.3
streamlit==1.39.0
swifter==1.4.0
sympy==1.12
systemd-python==234
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.20.3
toml==0.10.2
tomli==2.0.2
toolz==1.0.0
torch==2.4.1
torchcam==0.4.0
torchmetrics==1.3.2
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.46.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ua-parser==0.18.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
ujson==5.10.0
ultralytics==8.3.92
ultralytics-thop==2.0.14
umap-learn==0.5.6
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.20
usb-creator==0.3.7
user-agents==2.2.0
userspacefs==2.0.6
wadllib==1.3.3
wandb==0.21.0
watchdog==4.0.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wrapt==1.17.2
xkit==0.0.0
xmljson==0.2.1
xtcocotools==1.14.3
yapf==0.40.2
yarl==1.9.4
zipp==3.18.2
